Corrected tensor dimensions in the original model
4th and 7th layers pool temporal dimension by 2. Calculation
of padding produced tensor of larger than expected dimensions.

+: Correction of padding size.
+: Corrected implicit type casting to `torch.float32` in CE
   weighting of class imbalance.
+: Updates to config files.
+: Updates to README.

+++ processor.py
@@ -43,7 +43,7 @@ class Processor:
 
         for _, labels in dataloader:
             class_dist += torch.sum(
-                labels[:,:,None] == classes[None].expand(labels.shape[1],-1),
+                (labels[:,:,None].to(torch.float32) == classes[None].expand(labels.shape[1],-1)).to(torch.float32),
                 dim=(0,1))
 
         self.model = model
